Updating Alias tutorials to Drac2 (#1482)
* Update aliasing_tut.rst

* Update aliasing_tut.rst

* Update aliasing_tut.rst

* Update aliasing_tut.rst

Updating Orc-Relentless alias to <drac2></drac2> and including some if-blocks. Still need to update the explanation documentation

* Update aliasing_tut.rst

Complete overhaul of the documentation

* Update aliasing_tut.rst

Some formatting updates since it didn't like spaces in the `` `` blocks

* Update aliasing_tut.rst

Adding characters to the explanation as suggested.

* Update aliasing_tut.rst

Aaaand undoing that because it didn't look very good.

* Update aliasing_tut.rst

Adding bulleted lists as recommended.