AFR-220: Set character per server
Added commands to sheetManager:
!char server <name>: changes the character active on the current server, shows an error in DMs.
!char unset: removes the character as the active server character, defaulting to global.
!char global: functions the same a !char except if you put no name it will show the active global character.

Also made !char, !char server, and !char global show the active contexts (SERVER and/or GLOBAL) as well as the sheet type.

Added active_guilds as a local property of Character(), as well as incorporated it into the mdb to allow for searching for characters based on which servers they are active in.