Fix problems
I fixed the problems Zhu pointed out and added the confirmation for subscribing to workshop collections.  Also changed the flow of logic to make the declaration of variables more clear.

Sadly I don't know a way for me to test the serving of workshop collections at the time so this part has not been tested