Update custom !init attack docs to be clearer
Clarify the documentation on making one-off custom attacks to indicate that the attack needs to be named, the name can be anything, and "custom" is a modifier. The previous wording was unclear about whether "custom" was a modifier, or to be specified as the name of the attack; this hopefully improves that.